protocol/bc: new files for typed inputs
